+++ lib/rebit/models/resource_base.rb
@@ -1,7 +1,10 @@
 module Rebit
   class ResourceBase < ActiveResource::Base
 
-    self.site = "#{HOST}/api/#{API_VERSION}"
+    self.site = HOST
+    self.prefix = "/api/#{API_VERSION}/"
+    self.include_root_in_json = true
+    self.remove_root = true
 
   end
 end