+++ lib/rebit/models/vendor_user.rb
@@ -1,0 +1,7 @@
+module Rebit
+  class VendorUser < VendorBase
+
+    self.element_name = "user"
+
+  end
+end