+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # TBA
 
 - Add `#rates` call
+- Add VendorUser model for managing vendor users